Added service registration test of DI system for Web API (#1845)
* Added service registration test of DI system for Web API

* Added command line args to startup

* Fixed erroneous program setup

* Cleanup

* cleanup

* Revert static Program for web api
